Switch from atomic loads to memory fences in new hashtable (#266)
This PR changes how the data are read from the hashtable buckets,
switching from atomic loads to load memory fences and plain loads.

In the process of ensuring that there were no more weird sync bugs
between the threads the code was switched to use atomic loads but the
algorithm doesn't need them.
The code uses then an atomic compare & exchange to ensure that the value
that is going to be swapped is actually the value in memory, the code
will simply retry the operation as it needs to do anyway in case another
atomic compare & exchange is performed in between by another thread.

The difference between plain memory fences and loads VS atomic loads is
a lot, using load memory fences and plain loads provides a ~30%
performance gain.

The main reason is that an atomic operation, to ensure the atomicity,
does always a full memory barrier before and afterwards, sync between
all the CPU cores and numa nodes and then lock on the cacheline, .

But in the new hashtable this is stricly necessary, infact all the data
are set using atomic compare & exchange, which always ensure that the
value being written will replace the value that the code expects to be
there.

+++ src/data_structures/hashtable_mpmc/hashtable_mpmc.h
@@ -114,7 +114,7 @@ typedef struct hashtable_mpmc hashtable_mpmc_t;
 struct hashtable_mpmc {
     hashtable_mpmc_data_t *data;
     uint64_t buckets_count_max;
-    uint16_t upsize_preferred_block_size;
+    uint64_t upsize_preferred_block_size;
     hashtable_mpmc_upsize_info_t upsize;
 };
 
@@ -163,7 +163,7 @@ void hashtable_mpmc_data_free(
 hashtable_mpmc_t *hashtable_mpmc_init(
         uint64_t buckets_count,
         uint64_t buckets_count_max,
-        uint16_t upsize_preferred_block_size);
+        uint64_t upsize_preferred_block_size);
 
 void hashtable_mpmc_free(
         hashtable_mpmc_t *hashtable_mpmc);
